% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepMPT.R
\name{prep_Quad}
\alias{prep_Quad}
\title{Preparing Data for Quad Model}
\usage{
prep_Quad(data, Subject, Compatibility, Compatible = 1,
  Incompatible = 2, StimulusCat, Correct, TargetCat1, TargetCat2,
  AttributeCat1, AttributeCat2)
}
\arguments{
\item{data}{Dataframe containing all relevant columns (Subject, Compatibility, StimulusCat and Correct).}

\item{Subject}{Name of the column containing the subject id.}

\item{Compatibility}{Name of the column containing the compatibility of current key mapping.}

\item{Compatible}{Level of Compatibility coding a compatible mapping.}

\item{Incompatible}{Level of Compatibility coding an incompatible mapping.}

\item{StimulusCat}{Name of the column containing the category of the current stimulus.}

\item{Correct}{Name of the column containing accuracy of responses (0 = error, 1 = correct).}

\item{TargetCat1}{Level of StimulusCat coding the first target category (i.e, flower)}

\item{TargetCat2}{Level of StimulusCat coding the second target category (i.e, insect)}

\item{AttributeCat1}{Level of StimulusCat coding the first attribute category (i.e, pleasant)}

\item{AttributeCat2}{Level of StimulusCat coding the second attribute category (i.e, unpleasant)}
}
\description{
This function allows you to prepare IAT data for the application of functions estimating Quad model parameters.
}
\examples{
prep_Quad()
}
\keyword{IAT}
\keyword{Quad}
\keyword{mpt}
