% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepMPT.R
\name{prep_ReAL}
\alias{prep_ReAL}
\title{Preparing Data for ReAL Model estimation}
\usage{
prep_ReAL(data, Subject, TaskSwitch = NULL, Switch = 1,
  Repetition = 0, Trial = NULL, Compatibility, Compatible = 1,
  Incompatible = 2, StimulusCat, Correct, TargetCat1, TargetCat2,
  AttributeCat1, AttributeCat2)
}
\arguments{
\item{data}{Dataframe containing all relevant columns (Subject, TaskSwitch, Trial, Compatibility, StimulusCat and Correct).}

\item{Subject}{Name of the column containing the subject id.}

\item{TaskSwitch}{Name of the column containing infomation about task switching.}

\item{Switch}{Level of TaskSwitch coding task switchung.}

\item{Repetition}{Level of TaskSwitch coding task repetition.}

\item{Trial}{Name of the column containing the trial number.}

\item{Compatibility}{Name of the column containing the compatibility of current key mapping.}

\item{Compatible}{Level of Compatibility coding a compatible mapping.}

\item{Incompatible}{Level of Compatibility coding an incompatible mapping.}

\item{StimulusCat}{Name of the column containing the category of the current stimulus.}

\item{Correct}{Name of the column containing accuracy of responses (0 = error, 1 = correct).}

\item{TargetCat1}{Level of StimulusCat coding the first target category (i.e, flower)}

\item{TargetCat2}{Level of StimulusCat coding the second target category (i.e, insect)}

\item{AttributeCat1}{Level of StimulusCat coding the first attribute category (i.e, pleasant)}

\item{AttributeCat2}{Level of StimulusCat coding the second attribute category (i.e, unpleasant)}
}
\description{
This function allows you to prepare IAT data for the application of functions estimating ReAL model parameters.
}
\examples{
prep_ReAL()

}
\keyword{IAT}
\keyword{ReAL}
\keyword{mpt}
